requirements.txt
Pillow==9.0.1
pipdeptree==2.2.1
PyYAML==6.0
setuptools==58.1.0
tqdm==4.63.1
